Simply SparseBitVector includes.
