+++ svf/include/Util/SparseBitVector.h
@@ -16,6 +16,8 @@
 #include <iterator>
 #include <list>
 
+#include "Util/Hash.h"
+
 // Appease GCC?
 #ifdef __has_builtin
 #  define HAS_CLZ __has_builtin(__builtin_clz)
@@ -1244,5 +1246,16 @@ void dump(const SparseBitVector<ElementSize> &LHS, std::ostream &out)
 }
 
 } // End namespace SVF
+
+/// Specialise hash for SparseBitVectors.
+template <unsigned N> struct std::hash<SVF::SparseBitVector<N>>
+{
+    size_t operator()(const SVF::SparseBitVector<N>& sbv) const
+    {
+        SVF::Hash<std::pair<std::pair<size_t, size_t>, size_t>> h;
+        return h(std::make_pair(std::make_pair(sbv.count(), sbv.find_first()),
+                                sbv.find_last()));
+    }
+};
 
 #endif // SPARSEBITVECTOR_H
